Update patch set 1
Patch Set 1:

The shortcut from The link in 06-openstack-cli-next- should got to 
https://swift.rc.nectar.org.au/v1/AUTH_a404f16e313b4b09ac35d183a2f95afe/jekyll-test-build-I417c107d5eaa5708b5313827858fe3401a2ca0e0-1/cloud-two/
and does work in rake.
I did not change the structure just removed the non-existent page it was pointing to so I only had one place holder to create. All the other broken links only pointed to /cloud-two/

Patch-set: 1